refactor(ui): split chat state container
